event_driven/requirements.txt
flask
numpy
pandas
psycopg2
psutil
mpi4py
python-dotenv